backend/requirements.txt
openai
flask
flask-cors
python-dotenv
supabase
